+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules
 test/npm-debug.log
-test/server
+test/server*
 package-lock.json
 versions/
 src/client/*.json
