+++ src/client/authConstants.js
@@ -1,11 +1,6 @@
-const XSTSRelyingParty = 'rp://api.minecraftservices.com/'
-const MinecraftServicesLogWithXbox = 'https://api.minecraftservices.com/authentication/login_with_xbox'
-const MinecraftServicesEntitlement = 'https://api.minecraftservices.com/entitlements/mcstore'
-const MinecraftServicesProfile = 'https://api.minecraftservices.com/minecraft/profile'
-
 module.exports = {
-  XSTSRelyingParty,
-  MinecraftServicesLogWithXbox,
-  MinecraftServicesEntitlement,
-  MinecraftServicesProfile
+  XSTSRelyingParty: 'rp://api.minecraftservices.com/',
+  MinecraftServicesLogWithXbox: 'https://api.minecraftservices.com/authentication/login_with_xbox',
+  MinecraftServicesEntitlement: 'https://api.minecraftservices.com/entitlements/mcstore',
+  MinecraftServicesProfile: 'https://api.minecraftservices.com/minecraft/profile'
 }
